requirements.txt
# ============================================================================
# Athena Python Dependencies
# Production-grade requirements for ML-powered code intelligence
# ============================================================================

# ============================================================================
# Core Framework
# ============================================================================
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ============================================================================
# Database & ORM
# ============================================================================
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.0

# ============================================================================
# Message Queue
# ============================================================================
aiokafka==0.10.0
kafka-python==2.0.2

# ============================================================================
# GitHub API
# ============================================================================
PyGithub==2.1.1
aiohttp==3.9.1
httpx==0.26.0

# ============================================================================
# Caching
# ============================================================================
redis==5.0.1
hiredis==2.3.2

# ============================================================================
# Vector Database
# ============================================================================
weaviate-client==4.4.0

# ============================================================================
# ML/Deep Learning
# ============================================================================
# Core ML
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2

# Gradient Boosting
xgboost==2.0.3
lightgbm==4.1.0

# Deep Learning
torch==2.5.1
transformers==4.36.2
sentence-transformers==2.2.2

# Graph Neural Networks
torch-geometric==2.4.0
networkx==3.2.1

# Reinforcement Learning
gymnasium==0.29.1
stable-baselines3==2.2.1

# ============================================================================
# Feature Engineering
# ============================================================================
pyarrow==14.0.2
fastparquet==2023.10.1

# ============================================================================
# NLP & Code Analysis
# ============================================================================
tree-sitter==0.20.4
tree-sitter-python==0.21.0
tokenizers==0.15.0

# ============================================================================
# Experiment Tracking
# ============================================================================
mlflow==2.9.2

# ============================================================================
# Monitoring & Observability
# ============================================================================
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0

# ============================================================================
# Utilities
# ============================================================================
tenacity==8.2.3
loguru==0.7.2
click==8.1.7
tqdm==4.66.1
rich==13.7.0

# ============================================================================
# Data Validation
# ============================================================================
jsonschema==4.20.0
marshmallow==3.20.1

# ============================================================================
# Async Utilities
# ============================================================================
asyncio==3.4.3
aiofiles==23.2.1
aioboto3==12.3.0

# ============================================================================
# Testing
# ============================================================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==21.0.0

# ============================================================================
# Code Quality
# ============================================================================
black==23.12.1
flake8==7.0.0
mypy==1.7.1
pylint==3.0.3
isort==5.13.2

# ============================================================================
# Documentation
# ============================================================================
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# ============================================================================
# Development
# ============================================================================
ipython==8.18.1
jupyter==1.0.0
notebook==7.0.6
